More rewriting

+++ core/src/syn/parser/token_buffer.rs
@@ -29,7 +29,7 @@ impl<const S: usize> TokenBuffer<S> {
 
 	#[inline]
 	pub fn push_front(&mut self, token: Token) {
-		let next_read = self.read.checked_sub(1).unwrap_or(S - 1);
+		let next_read = self.read.checked_sub(1).unwrap_or((S - 1) as u8);
 		if next_read == self.write {
 			panic!("token buffer full");
 		}
